+++ README.md
@@ -1,19 +1,21 @@
 # Azure Notification Hubs for Go(lang)
 
 This library provides a Microsoft Azure Notification Hubs Client for backend applications.
-It is packaged as a Go module to and is tested with Go 1.11+.
+It is packaged as a Go module to and is tested with Go 1.12+.
 
 Originally a fork from [Gozure](https://github.com/onefootball/gozure) with patches
-from [Martin Etnestad](https://github.com/gnawybol).
+from [Martin Etnestad](https://github.com/gnawybol) @ [vippsas](https://github.com/vippsas/gozure).
 
 Now maintained and packaged by [Daresay AB](https://daresay.co), [@daresaydigital](https://github.com/daresaydigital).
+
+Basically a wrapper for this [Rest API](https://docs.microsoft.com/en-us/rest/api/notificationhubs/rest-api-methods)
 
 ## Installing
 
 Using go get
 
 ```sh
-$ go get github.com/daresaydigital/azure-notificationhubs-go
+go get github.com/daresaydigital/azure-notificationhubs-go
 ```
 
 ## Usage
@@ -22,41 +24,47 @@ $ go get github.com/daresaydigital/azure-notificationhubs-go
 package notificationhubs
 
 import (
-    "fmt"
-    "github.com/daresaydigital/azure-notificationhubs-go"
+  "fmt"
+  "github.com/daresaydigital/azure-notificationhubs-go"
 )
 
 func main() {
-    payload := []byte(`{"title": "Hello Hub!"}`)
+  var (
+    hub = notificationhubs.NewNotificationHub("YOUR_DefaultFullSharedAccessConnectionString", "YOUR_HubPath")
+    payload = []byte(`{"title": "Hello Hub!"}`)
+    n = notificationhubs.NewNotification(notificationhubs.Template, payload)
+  )
 
-    n, err := notificationhubs.NewNotification(notificationhubs.Template, payload)
-    if err != nil {
-        panic(err)
-    }
+  // broadcast push
+  b, err := hub.Send(n, nil)
+  if err != nil {
+    panic(err)
+  }
 
-    hub := notificationhubs.NewNotificationHub("YOUR_DefaultFullSharedAccessConnectionString", "YOUR_HubPath")
+  fmt.Println("Message successfully created:", string(b))
 
-    // broadcast push
-    b, err := hub.Send(n, nil)
-    if err != nil {
-        panic(err)
-    }
+  // tag category push
+  b, err = hub.Send(n, []string{"tag1", "tag2"})
+  if err != nil {
+    panic(err)
+  }
 
-    fmt.Println("message successfully created:", string(b))
-
-    // tag category push
-    b, err = hub.Send(n, []string{"tag1", "tag2"})
-    if err != nil {
-        panic(err)
-    }
-
-    fmt.Println("message successfully created:", string(b))
+  fmt.Println("Message successfully created:", string(b))
 }
 ```
 
 ## Changelog
 
+### v0.0.2
+
+- Big rewrite
+- Added get registrations
+- Travis CI
+- Renamed the entities to use the same nomenclature as Azure
+- Using fixtures for tests
+
 ### v0.0.1
+
 First release by Daresay. Restructured the code and renamed the API according to
 Go standards.
